+++ packages/polaris-core/src/config/polaris-server-config.ts
@@ -7,6 +7,7 @@ import { DocumentNode } from 'graphql';
 import { IResolvers } from 'graphql-tools';
 import { ExpressContext } from '..';
 import { MiddlewareConfiguration } from '../index';
+import { PermissionsConfiguration } from './permissions-configuration';
 import { SnapshotConfiguration } from './snapshot-configuration';
 
 export interface PolarisServerConfig extends Omit<ApolloServerExpressConfig, 'logger'> {
@@ -25,4 +26,5 @@ export interface PolarisServerConfig extends Omit<ApolloServerExpressConfig, 'lo
     snapshotConfig: SnapshotConfiguration;
     connectionManager?: PolarisConnectionManager;
     enableFederation: boolean;
+    permissionsConfig: PermissionsConfiguration;
 }