+++ packages/polaris-schema/src/main.ts
@@ -1,8 +1,8 @@
 export { RepositoryEntity } from './common/repository-entity';
 export { repositoryEntityTypeDefs } from './common/repository-entity-type-defs';
+export { PermissionsDirective } from './directives/permissions-directive';
 export { scalarsResolvers } from './scalars/scalars-resolvers';
 export { scalarsTypeDefs } from './scalars/scalars-type-defs';
 export { makeExecutablePolarisSchema } from './utils/executable-schema-creator';
 export { getMergedPolarisTypes } from './utils/merge-types';
 export { getMergedPolarisResolvers } from './utils/merge-resolvers';
-export { SchemaDirectiveVisitor } from 'graphql-tools';