+++ packages/polaris-test/tests/whoami.test.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import * as polarisProperties from '../server-without-connection/resources/polaris-properties.json';
+import * as polarisProperties from '../shared-resources/polaris-properties.json';
 import { createServersWithoutConnection } from '../test-utils/tests-servers-util';
 
 describe('whoami tests', () => {