+++ packages/polaris-graphql-logger/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.3](https://github.com/Enigmatis/polaris-graphql-logger/compare/@enigmatis/polaris-graphql-logger@2.0.2...@enigmatis/polaris-graphql-logger@2.0.3) (2020-09-09)
+
+**Note:** Version bump only for package @enigmatis/polaris-graphql-logger
+
+
+
+
+
 ## [2.0.2](https://github.com/Enigmatis/polaris-graphql-logger/compare/@enigmatis/polaris-graphql-logger@2.0.1...@enigmatis/polaris-graphql-logger@2.0.2) (2020-09-07)