+++ packages/polaris-schema/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/Enigmatis/polaris-schema/compare/@enigmatis/polaris-schema@1.3.0...@enigmatis/polaris-schema@1.3.1) (2020-09-10)
+
+
+### Bug Fixes
+
+* fixed permission directive definition ([#13](https://github.com/Enigmatis/polaris-schema/issues/13)) ([001f1bb](https://github.com/Enigmatis/polaris-schema/commit/001f1bb45c8c88fe4fef5c9308e9f35375d59245))
+
+
+
+
+
+# [1.3.0](https://github.com/Enigmatis/polaris-schema/compare/@enigmatis/polaris-schema@1.2.6...@enigmatis/polaris-schema@1.3.0) (2020-09-09)
+
+
+### Features
+
+* added permission directive ([#12](https://github.com/Enigmatis/polaris-schema/issues/12)) ([0b59db0](https://github.com/Enigmatis/polaris-schema/commit/0b59db056e40e1e417f932d02a21023c007ea5f9))
+
+
+
+
+
 ## [1.2.6](https://github.com/Enigmatis/polaris-schema/compare/@enigmatis/polaris-schema@1.2.5...@enigmatis/polaris-schema@1.2.6) (2020-09-07)